+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Added documentation for running a subset of the backend tests
 - Deleted `app/Response/Image` file and refactored `app/Util/Helpers` to include `returnImageData`
 - Added test cases for the retrieveProgram function in `Controllers/Program`
+- Removed duplicate code from `mockRequest` and `runServerPartWith` in `backend-test/TestHelpers.hs`
 - Initialized a SchemaVersion table for the purposes of running robust database migrations
 
 ## [0.7.1] - 2025-06-16